sam/classifier/requirements.txt
awslambdaric==2.0.4
Pillow==9.2.0
